+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## main (unreleased)
 
 - [#16](https://github.com/clojure-emacs/clojure-ts-mode/issues/16): Introduce `clojure-ts-align`.
+- [#11](https://github.com/clojure-emacs/clojure-ts-mode/issues/11): Enable regex syntax highlighting.
 
 ## 0.3.0 (2025-04-15)